fix(front): rename contant.js and small code improvements
