+++ frontend_lib/src/TracimTableLib/RowWrappers/ContentRowWrapper.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { Link } from 'react-router-dom'
 
-import { PAGE } from '../../helperConstants.js'
+import { PAGE } from '../../constant.js'
 import ListItemWrapper from '../../component/Lists/ListItemWrapper/ListItemWrapper.jsx'
 
 const ContentRowWrapper = (props) => {
