+++ frontend_lib/src/component/Layout/PageTitle.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { translate } from 'react-i18next'
 import classnames from 'classnames'
-import { BREADCRUMBS_TYPE, PAGE } from '../../helperConstants.js'
+import { BREADCRUMBS_TYPE, PAGE } from '../../constant.js'
 import Breadcrumbs from '../Breadcrumbs/Breadcrumbs.jsx'
 import Popover from '../Popover/Popover.jsx'
 import ComposedIcon from '../Icon/ComposedIcon.jsx'
