+++ frontend_lib/src/component/PopinFixed/PopinFixedContent.jsx
@@ -6,7 +6,7 @@ import SelectStatus from '../Input/SelectStatus/SelectStatus.jsx'
 import {
   APP_FEATURE_MODE,
   ROLE
-} from '../../helperConstants.js'
+} from '../../constant.js'
 import TranslateButton from '../Button/TranslateButton.jsx'
 import PopinFixedHeader from './PopinFixedHeader.jsx'
 import { TRANSLATION_STATE } from '../../translation.js'
