+++ frontend_lib/src/component/ProfileNavigation/ProfileNavigation.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import { Link } from 'react-router-dom'
-import { PAGE } from '../../helperConstants.js'
+import { PAGE } from '../../constant.js'
 import { translate } from 'react-i18next'
 
 const ProfileNavigation = props => {
