+++ frontend_lib/test/fixture/contentCommentList.js
@@ -1,5 +1,5 @@
 import { author } from './author.js'
-import { CONTENT_TYPE, TIMELINE_TYPE } from '../../src/helperConstants.js'
+import { CONTENT_TYPE, TIMELINE_TYPE } from '../../src/constant.js'
 
 export const commentList = [{
   author: author,
